Project2/requirements.txt
flask==3.0.0
boto3==1.34.0
pymongo==4.6.1
python-dotenv==1.0.0
werkzeug==3.0.1
Pillow==10.2.0
gunicorn==21.2.0
